requirements.txt
asyncio
psycopg2-binary
aiopg
nats-py
